+++ seq_add.c
@@ -2,13 +2,15 @@
 #include <math.h>
 #include <stdlib.h>
 #include <stdbool.h>
+#include<stdint.h>
+
 #include <math.h>
 #include<string.h>
 #include<fcntl.h>
-void print_results(char* name,bool * primes_bool,unsigned long long int min,unsigned long long int max){
+void print_results(char* name,bool * primes_bool,uint64_t min,uint64_t max){
     FILE* x=fopen(name,"w+");
     fclose(x);   
-    for (unsigned long long int i = min;i<=max;i++){
+    for (uint64_t i = min;i<=max;i++){
     FILE* f=fopen(name,"a");    
        
         
@@ -19,21 +21,21 @@ void print_results(char* name,bool * primes_bool,unsigned long long int min,unsi
     }}
  int main( int argc, char *argv[])
 {
-    unsigned long long int min = 2;
-    unsigned long long int max = 100000000;
-    unsigned long long int size = max - min;
+    uint64_t min = 2;
+    uint64_t max = 100000000;
+    uint64_t size = max - min;
     FILE f;
     bool prime = false;
-    unsigned long long int *primes;
+    uint64_t *primes;
     primes = malloc(sizeof(int) * size);
-    unsigned long long int num_primes = 0;
+    uint64_t num_primes = 0;
     bool * primes_bool;
     primes_bool=malloc(sizeof(bool)*(max + 1));
     //memset(primes_bool, true, sizeof(primes_bool))false;
 
-    unsigned long long int sqrt_sieve = sqrt(max);
-    unsigned long long int tmp; // temporary sum
-    for (unsigned long long int i = 2; i <= sqrt_sieve; i++)
+    uint64_t sqrt_sieve = sqrt(max);
+    uint64_t tmp; // temporary sum
+    for (uint64_t i = 2; i <= sqrt_sieve; i++)
     {
         if (primes_bool[i] == false)
         {
